more modular structure of js files
