+++ app/js/controllers/dropController.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = exports = (app) => {
+  app.controller('dropCtrl', function($scope) {
+    $scope.countrylist = ['United States', 'Canada', 'South Africa'];
+  });
+};
