+++ app/js/controllers/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+module.exports = function(app) {
+  require('./dropController')(app);
+};
